# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Word Search API",
        "description": "API documentation for Laravel Word Search application, including both internal and external endpoints.",
        "contact": {
            "name": "API Support",
            "email": "support@fairladymedia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://wordlists.fairladymedia.com",
            "description": "Production Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/v1/longest-word": {
            "get": {
                "tags": [
                    "Longest Word"
                ],
                "summary": "Get current longest word",
                "description": "Retrieves the longest word stored for the current session. Each player sees only their own longest word and receives their unique player_id.",
                "operationId": "getLongestWord",
                "responses": {
                    "200": {
                        "description": "Current longest word information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "longest_word": {
                                            "description": "The player's longest word, null if no words submitted",
                                            "type": "string",
                                            "example": "extraordinary",
                                            "nullable": true
                                        },
                                        "length": {
                                            "description": "Length of the longest word, 0 if no words submitted",
                                            "type": "integer",
                                            "example": 13
                                        },
                                        "player_id": {
                                            "description": "Unique identifier for the player (first 8 characters of session ID)",
                                            "type": "string",
                                            "example": "8f7d9c2e"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded (60 per minute)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Longest Word"
                ],
                "summary": "Submit a new word",
                "description": "Attempts to store a word if it's longer than the current longest word for the session. Each player maintains their own longest word record, identified by their session.",
                "operationId": "storeLongestWord",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "word"
                                ],
                                "properties": {
                                    "word": {
                                        "description": "The word to submit",
                                        "type": "string",
                                        "example": "extraordinary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Word submission result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_longest": {
                                            "description": "Whether this word became the player's new longest word",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "submitted_word": {
                                            "description": "The word that was submitted",
                                            "type": "string",
                                            "example": "extraordinary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - empty or invalid word"
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded (60 per minute)"
                    }
                }
            }
        },
        "/api/v1/longest-word/top": {
            "get": {
                "tags": [
                    "Longest Word"
                ],
                "summary": "Get top 10 longest words",
                "description": "Retrieves the top 10 longest words submitted by all players, ordered by word length. Each word includes the submitter's player_id for identification.",
                "operationId": "getTopLongestWords",
                "responses": {
                    "200": {
                        "description": "Top 10 longest words",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "words": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "word": {
                                                        "description": "The submitted word",
                                                        "type": "string",
                                                        "example": "supercalifragilistic"
                                                    },
                                                    "player_id": {
                                                        "description": "Unique identifier for the player who submitted the word",
                                                        "type": "string",
                                                        "example": "8f7d9c2e"
                                                    },
                                                    "length": {
                                                        "description": "Length of the word",
                                                        "type": "integer",
                                                        "example": 20
                                                    },
                                                    "submitted_at": {
                                                        "description": "When the word was submitted",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-15T10:30:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded (60 per minute)"
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Word Search"
                ],
                "summary": "Display the word search interface",
                "operationId": "59f0741d68535f9065795aab9930413a",
                "responses": {
                    "200": {
                        "description": "Word search view"
                    }
                }
            }
        },
        "/search": {
            "post": {
                "tags": [
                    "Word Search"
                ],
                "summary": "Search for words based on pattern",
                "description": "Search through specialized word lists based on a query pattern. Protected by CSRF token.",
                "operationId": "8896c7f75118ec3ac626e94b58fcf295",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query",
                                    "list"
                                ],
                                "properties": {
                                    "query": {
                                        "description": "The search pattern",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "list": {
                                        "description": "Which list to search in",
                                        "type": "string",
                                        "enum": [
                                            "omnigrams",
                                            "wordchecker",
                                            "both"
                                        ],
                                        "example": "both"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "results": {
                                            "properties": {
                                                "omnigrams": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "wordchecker": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/search-frequency": {
            "post": {
                "tags": [
                    "Word Search"
                ],
                "summary": "Search for words based on frequency",
                "description": "Search for words that meet or exceed a specified frequency threshold. Protected by CSRF token.",
                "operationId": "08a95d20d9a10a056460690268b92cff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "frequency"
                                ],
                                "properties": {
                                    "frequency": {
                                        "description": "Minimum frequency threshold",
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.0e-7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "total_count": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "words": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "csrf": {
                "type": "apiKey",
                "description": "CSRF token required for internal API endpoints",
                "name": "X-CSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Longest Word",
            "description": "External API endpoints for tracking longest words found by players. Each player is identified by their session, with a shortened player_id for display."
        },
        {
            "name": "Word Search",
            "description": "Internal API endpoints for word searching functionality"
        }
    ],
    "externalDocs": {
        "description": "Find more info in README",
        "url": "https://github.com/yourusername/laravel-word-search"
    },
    "security": [
        {
            "csrf": []
        }
    ]
}